Refactors code to improve readability
Changes split on /create_table/ to /create_/ allowing to use the
'table' word to make it easier to scan for the table name.
Changes open file method to properly close the file after reading.
Adds a block to String#scan call using the captures passed to it to
print the options.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>