Complete remake and restructure to gemify
Add Schema, Table and Attribute classes to the new Schemaffold Module.
Restructure code so that it is possible to distribute a linux command
line application.
The command line app accepts the path to the schema.rb file so that
the user no longer needs to look inside the code.
Add copy script to clipboard feature using xclip in linux

Signed-off-by: João Soares <jsoaresgeral@gmail.com>